backend/dns-service/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
gevent==21.8.0
python-dotenv==0.19.0
pytest==6.2.5
dnspython==2.1.0
